Refactored code to separate offersReceived and offersSent components
