+++ src/components/Offering/OffersReceived/DeclineIconButton.js
@@ -3,7 +3,7 @@ import { makeStyles } from "@material-ui/core/styles";
 import { IconButton } from "@material-ui/core";
 import HighlightOffIcon from "@material-ui/icons/HighlightOff";
 
-import { HoverPopoverHOC } from "../HigherOrderComponents/HoverPopoverHOC";
+import { HoverPopoverHOC } from "../../HigherOrderComponents/HoverPopoverHOC";
 
 const useStyles = makeStyles(() => ({
   actionBtnIcon: {
