+++ src/components/Offering/OffersReceived/OfferInfoDisplay.js
