+++ src/components/Offering/OffersReceived/OffererInfoDisplay.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Grid, Typography } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 
-import UserAvatar from "../User/UserAvatar";
+import UserAvatar from "../../User/UserAvatar";
 
 const useStyles = makeStyles(() => ({
   infoContainer: {
