+++ src/components/Offering/OffersReceived/OfferingDetails.js
@@ -5,8 +5,8 @@ import { makeStyles } from "@material-ui/core/styles";
 
 import { OffererInfoDisplay } from "./OffererInfoDisplay";
 import { OfferInfoDisplay } from "./OfferInfoDisplay";
-import { closeModal } from "../../redux/actions/modalActions";
-import { acceptOffer, declineOffer } from "../../redux/actions/offeringActions";
+import { closeModal } from "../../../redux/actions/modalActions";
+import { acceptOffer, declineOffer } from "../../../redux/actions/offeringActions";
 
 let useStyles = makeStyles((theme) => ({
   paper: {
