+++ src/components/Offering/OffersReceived/OfferingHelpers.js
@@ -1,5 +1,5 @@
 import _ from "lodash";
-import { offeringStatus } from "../constants/OfferingConstants";
+import { offeringStatus } from "../../constants/OfferingConstants";
 
 const addNumPendingOffers = (sum, activePosting) => {
   let pendingOffers = getAllPendingOffers(activePosting.offerings);
