+++ src/components/Offering/OffersReceived/OfferingPreviewList.js
