+++ src/components/User/UserDetails.js
@@ -9,9 +9,9 @@ import TabLabel from "./TabLabel";
 import ReviewList from "../Review/ReviewList";
 import ItemPreviewList from "../Item/ItemPreviewList";
 import Button from "@material-ui/core/Button";
-import { calculateTotalPendingOffersReceived } from "../Offering/OfferingHelpers";
-import { OffersReceived } from "../Offering/OffersReceived";
-import { OffersSent } from "../Offering/OffersSent";
+import { calculateTotalPendingOffersReceived } from "../Offering/OffersReceived/OfferingHelpers";
+import { OffersReceived } from "../Offering/OffersReceived/OffersReceived";
+import { OffersSent } from "../Offering/OffersSent/OffersSent";
 
 const useStyles = (theme) => ({
   root: {
